scripts/requirements.txt
pyyaml
beautifulsoup4

